+++ exp/pecos-frontier/src/lib.rs
@@ -10,7 +10,7 @@
 // or implied. See the License for the specific language governing permissions and limitations under
 // the License.
 
-//! Trellis-class approximate logical maximum-likelihood decoding.
+//! Frontier approximate logical maximum-likelihood decoding.
 //!
 //! The decoder performs ordered dynamic programming over independent binary
 //! fault mechanisms. Prefixes with identical active detector boundary and
@@ -18,18 +18,8 @@
 //! configured frontier width and log-mass window provide deterministic pruning
 //! for a fixed build and platform; underlying `ln`/`exp` implementations may
 //! differ across platforms.
-//!
-//! [`BpTrellisDecoder`] is PECOS's own degeneracy-aware, coset-mass decoder in
-//! this class. It is exact in the unpruned limit; with pruning it is
-//! approximate and has no certified bound on discarded posterior mass. Its
-//! optimality is relative to the supplied detector error model (DEM), not the
-//! underlying physics. Belief propagation (BP) guides only which states
-//! pruning retains and never changes branch probabilities or mass arithmetic.
-//! It is not a wrap or port of an external project.
-//!
-//! In contrast, [`FrontierDecoder`] is the parity port of Leverrier and
-//! Urbanke's Frontier decoder. Its input-order defaults remain those of that
-//! port. Both types use the same engine and return [`FrontierResult`].
+//! [`FrontierDecoder`] is the parity port of Leverrier and Urbanke's Frontier
+//! decoder. Its input-order defaults remain those of that port.
 
 use pecos_decoder_core::ObservableDecoder;
 use pecos_decoder_core::bp::{BpGraph, BpScratch, min_sum_bp_into};
@@ -85,67 +75,6 @@ impl Default for FrontierConfig {
             column_order: None,
             merge_indistinguishable: false,
             bp_score_iterations: 0,
-        }
-    }
-}
-
-/// Processing order used by [`BpTrellisDecoder`].
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub enum TrellisOrdering {
-    /// Compute the deadline-optimized order with [`deadline_column_order`].
-    #[default]
-    Deadline,
-    /// Compute the backward deadline-optimized order with
-    /// [`backward_deadline_column_order`].
-    BackwardDeadline,
-    /// Preserve the detector error model's mechanism order.
-    TimeOrder,
-    /// Use an explicit permutation mapping target positions to source
-    /// mechanism indices.
-    Explicit(Vec<usize>),
-}
-
-/// Configuration for PECOS's [`BpTrellisDecoder`].
-///
-/// These defaults are provisional. In particular, `k = 8` was validated as
-/// near-floor on the BB144 benchmark used to select BP-guided retention, but
-/// broader code and noise-model validation is still pending.
-#[derive(Clone, Debug, PartialEq)]
-pub struct BpTrellisConfig {
-    /// Maximum number of boundary states retained after each column.
-    pub k: usize,
-    /// Log-mass window below the best boundary state retained after each
-    /// column.
-    pub delta: f64,
-    /// Weight applied to the suffix-compatibility score during pruning.
-    pub score_alpha: f64,
-    /// Number of min-sum BP iterations used only to score pruning candidates.
-    pub bp_score_iterations: usize,
-    /// Merge probabilistic mechanisms with identical detector and observable
-    /// sets using their exact XOR-combined probability.
-    pub merge_indistinguishable: bool,
-    /// Mechanism processing order.
-    pub ordering: TrellisOrdering,
-    /// Escalation ladder used only after a no-path decode.
-    ///
-    /// Each entry pre-builds another decoder with that `k` and otherwise the
-    /// same configuration. Construction cost therefore grows with the number
-    /// of rungs. An empty ladder disables escalation and is the explicit
-    /// default because escalation changes per-shot work; whether a future
-    /// default should enable a ladder is deferred to the evaluation campaign.
-    pub escalation_ks: Vec<usize>,
-}
-
-impl Default for BpTrellisConfig {
-    fn default() -> Self {
-        Self {
-            k: 8,
-            delta: 100.0,
-            score_alpha: 0.8,
-            bp_score_iterations: 5,
-            merge_indistinguishable: true,
-            ordering: TrellisOrdering::Deadline,
-            escalation_ks: Vec::new(),
         }
     }
 }
@@ -235,7 +164,7 @@ pub struct FrontierResult {
     pub processed_columns: usize,
     /// Number of candidate branch evaluations, counted at entry to
     /// `merge_branch` (two per retained state for every processed column).
-    /// For an escalated [`BpTrellisDecoder`] result, this is the total across
+    /// For an escalated `BpTrellis` result, this is the total across
     /// the base attempt and every attempted rung.
     pub transitions: u64,
     /// Number of merged boundary states discarded across all pruning calls in
@@ -247,23 +176,22 @@ pub struct FrontierResult {
     /// This accounts for retained prefix mass discarded at pruning time. It is
     /// not a bound on true lost posterior mass: a state dropped early would
     /// otherwise have branched through later columns.
-    /// For an escalated [`BpTrellisDecoder`] result, this covers only the
+    /// For an escalated `BpTrellis` result, this covers only the
     /// successful rung.
     pub dropped_log_mass: f64,
     /// Wall-clock seconds spent producing BP-informed suffix scores for this
     /// shot. This is zero when BP scoring is disabled or pruning cannot run.
-    /// For an escalated [`BpTrellisDecoder`] result, this is the total across
+    /// For an escalated `BpTrellis` result, this is the total across
     /// the base attempt and every attempted rung.
     pub bp_seconds: f64,
     /// Number of escalation rungs attempted before success.
     ///
     /// Zero means the base decode succeeded. [`FrontierDecoder`] results and
-    /// successful decodes from a ladder-free [`BpTrellisDecoder`] also report
-    /// zero.
+    /// successful ladder-free `BpTrellis` decodes also report zero.
     pub escalation_rungs_used: u32,
     /// Whether the successful decode was exact or which pruning mechanisms
-    /// discarded at least one state. For an escalated [`BpTrellisDecoder`]
-    /// result, this is the successful rung's status.
+    /// discarded at least one state. For an escalated `BpTrellis` result, this
+    /// is the successful rung's status.
     pub status: FrontierStatus,
     /// Retained unnormalized joint terminal masses, ordered by mass descending
     /// and numeric label ascending. The first entry is the winning label and
@@ -395,13 +323,23 @@ type RawColumn = (Vec<u64>, Vec<u64>, f64);
 type SuffixCompatibilityTables = Vec<Vec<SuffixCompatibility>>;
 type BpSuffixPreparation = (Option<SuffixCompatibilityTables>, f64);
 
-enum FrontierDecodeAttempt {
+/// Structured outcome of one Frontier engine attempt.
+///
+/// This preserves work telemetry for a no-path outcome so higher-level decode
+/// policies can retry without duplicating engine logic.
+pub enum FrontierDecodeAttempt {
+    /// The attempt retained at least one terminal state.
     Success(FrontierResult),
+    /// The attempt retained no path for the observed syndrome.
     NoPath {
+        /// Error reported when no higher-level retry succeeds.
         error: DecoderError,
+        /// Candidate branch evaluations performed by this attempt.
         transitions: u64,
+        /// Wall-clock seconds spent on BP suffix scoring in this attempt.
         bp_seconds: f64,
     },
+    /// A non-retryable decoding error.
     Error(DecoderError),
 }
 
@@ -426,24 +364,6 @@ pub struct FrontierDecoder {
     forced_syndrome: Vec<u64>,
     forced_logical: Vec<u64>,
     bp_score: Option<BpScoreState>,
-    build_seconds: f64,
-}
-
-/// PECOS's BP-guided trellis decoder for logical coset posterior mass.
-///
-/// This is a degeneracy-aware approximate logical maximum-likelihood decoder,
-/// exact in the unpruned limit. It is optimal relative to the supplied DEM,
-/// not the underlying physics. Pruned results have no certified bound on
-/// discarded posterior mass. BP guides only which states pruning retains and
-/// never changes the engine's branch probabilities or mass arithmetic.
-///
-/// This PECOS decoder is not a wrap or port of an external project. It owns a
-/// [`FrontierDecoder`] configured with PECOS's defaults and ordering semantics
-/// and shares that engine's [`FrontierResult`].
-#[derive(Clone, Debug)]
-pub struct BpTrellisDecoder {
-    inner: FrontierDecoder,
-    escalation: Vec<FrontierDecoder>,
     build_seconds: f64,
 }
 
@@ -667,7 +587,10 @@ impl FrontierDecoder {
         self.decode_attempt(syndrome).into_result()
     }
 
-    fn decode_attempt(&mut self, syndrome: &[u8]) -> FrontierDecodeAttempt {
+    /// Decode while preserving no-path work telemetry for higher-level retry
+    /// policies.
+    #[must_use]
+    pub fn decode_attempt(&mut self, syndrome: &[u8]) -> FrontierDecodeAttempt {
         if syndrome.len() != self.num_detectors {
             return FrontierDecodeAttempt::Error(DecoderError::InvalidDimensions {
                 expected: self.num_detectors,
@@ -853,141 +776,6 @@ impl FrontierDecoder {
     }
 }
 
-impl BpTrellisDecoder {
-    /// Construct a decoder from a sparse detector error model.
-    ///
-    /// Unlike [`FrontierDecoder`], the default ordering is the explicitly
-    /// computed [`deadline_column_order`], not input order. Every configured
-    /// escalation rung is constructed here as an independent
-    /// [`FrontierDecoder`], so construction cost scales with the full ladder
-    /// and decode-time escalation performs no model building.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`DecoderError`] if ordering generation or the mapped Frontier
-    /// configuration fails validation.
-    pub fn from_sparse_dem(dem: &SparseDem, config: BpTrellisConfig) -> Result<Self, DecoderError> {
-        let build_started = Instant::now();
-        let BpTrellisConfig {
-            k,
-            delta,
-            score_alpha,
-            bp_score_iterations,
-            merge_indistinguishable,
-            ordering,
-            escalation_ks,
-        } = config;
-        if u32::try_from(escalation_ks.len()).is_err() {
-            return Err(DecoderError::InvalidConfiguration(
-                "escalation ladder has more rungs than escalation_rungs_used can represent".into(),
-            ));
-        }
-        let column_order = match ordering {
-            TrellisOrdering::Deadline => Some(deadline_column_order(dem)?),
-            TrellisOrdering::BackwardDeadline => Some(backward_deadline_column_order(dem)?),
-            TrellisOrdering::TimeOrder => None,
-            TrellisOrdering::Explicit(order) => Some(order),
-        };
-        let frontier_config = FrontierConfig {
-            k,
-            delta,
-            score_alpha,
-            column_order,
-            merge_indistinguishable,
-            bp_score_iterations,
-        };
-        let inner = FrontierDecoder::from_sparse_dem(dem, frontier_config.clone())?;
-        let escalation = escalation_ks
-            .into_iter()
-            .map(|rung_k| {
-                FrontierDecoder::from_sparse_dem(
-                    dem,
-                    FrontierConfig {
-                        k: rung_k,
-                        ..frontier_config.clone()
-                    },
-                )
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        let build_seconds = build_started.elapsed().as_secs_f64();
-        Ok(Self {
-            inner,
-            escalation,
-            build_seconds,
-        })
-    }
-
-    /// Parse a Stim-format detector error model and construct a decoder.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`DecoderError`] if parsing, ordering generation, or decoder
-    /// validation fails.
-    pub fn from_dem_str(dem_str: &str, config: BpTrellisConfig) -> Result<Self, DecoderError> {
-        let dem = SparseDem::from_dem_str(dem_str)?;
-        Self::from_sparse_dem(&dem, config)
-    }
-
-    /// Decode a dense detector syndrome with the shared trellis engine.
-    ///
-    /// Every nonzero byte is treated as a fired detector. A no-path base
-    /// attempt is retried at each pre-built escalation rung until one
-    /// succeeds. Any returned prediction, including a pruned or incorrect
-    /// prediction, is final because production decoding has no truth oracle.
-    /// Other errors do not escalate. On success, `transitions` and
-    /// `bp_seconds` cover the full attempted sequence; pruning telemetry and
-    /// status come from the successful rung.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`DecoderError`] for a dimension mismatch or when the syndrome
-    /// is unexplainable with the retained frontier.
-    pub fn decode(&mut self, syndrome: &[u8]) -> Result<FrontierResult, DecoderError> {
-        let (mut final_error, mut transitions, mut bp_seconds) =
-            match self.inner.decode_attempt(syndrome) {
-                FrontierDecodeAttempt::Success(result) => return Ok(result),
-                FrontierDecodeAttempt::NoPath {
-                    error,
-                    transitions,
-                    bp_seconds,
-                } => (error, transitions, bp_seconds),
-                FrontierDecodeAttempt::Error(error) => return Err(error),
-            };
-
-        let mut escalation_rungs_used = 0_u32;
-        for decoder in &mut self.escalation {
-            escalation_rungs_used = escalation_rungs_used.saturating_add(1);
-            match decoder.decode_attempt(syndrome) {
-                FrontierDecodeAttempt::Success(mut result) => {
-                    result.transitions += transitions;
-                    result.bp_seconds += bp_seconds;
-                    result.escalation_rungs_used = escalation_rungs_used;
-                    return Ok(result);
-                }
-                FrontierDecodeAttempt::NoPath {
-                    error,
-                    transitions: rung_transitions,
-                    bp_seconds: rung_bp_seconds,
-                } => {
-                    final_error = error;
-                    transitions += rung_transitions;
-                    bp_seconds += rung_bp_seconds;
-                }
-                FrontierDecodeAttempt::Error(error) => return Err(error),
-            }
-        }
-
-        Err(final_error)
-    }
-
-    /// Total wall-clock seconds spent constructing the base decoder and every
-    /// escalation-rung model.
-    #[must_use]
-    pub fn build_seconds(&self) -> f64 {
-        self.build_seconds
-    }
-}
-
 impl FrontierCommittee {
     /// Construct a forward/backward committee from a sparse DEM.
     ///
@@ -1100,22 +888,6 @@ impl ObservableDecoder for FrontierDecoder {
 
     fn decode_to_observables(&mut self, syndrome: &[u8]) -> Result<u64, DecoderError> {
         if self.logical_words > 1 {
-            return Err(DecoderError::InvalidConfiguration(
-                "decoder has more than 64 observables; use decode_obs() for the wide mask".into(),
-            ));
-        }
-        let decoded = self.decode(syndrome)?.predicted;
-        Ok(decoded.words().first().copied().unwrap_or(0))
-    }
-}
-
-impl ObservableDecoder for BpTrellisDecoder {
-    fn decode_obs(&mut self, syndrome: &[u8]) -> Result<ObsMask, DecoderError> {
-        Ok(self.decode(syndrome)?.predicted)
-    }
-
-    fn decode_to_observables(&mut self, syndrome: &[u8]) -> Result<u64, DecoderError> {
-        if self.inner.logical_words > 1 {
             return Err(DecoderError::InvalidConfiguration(
                 "decoder has more than 64 observables; use decode_obs() for the wide mask".into(),
             ));
