+++ exp/pecos-frontier/tests/MUTANTS.md
@@ -26,9 +26,3 @@ row. “Equivalent” rows document mutations that intentionally have no killer.
 | `bp_tables_from_dem_priors` | In `bp_suffix_compatibility`, replace the call that builds tables from the BP-derived `moments` with clones of each column's static `suffix_compatibility` table. | `bp_scores_change_the_greedy_survivor_without_changing_its_mass` |
 | `bp_clamp_removed` | In `bp_score_probability`, replace the complete `probability.clamp(BP_SCORE_PROBABILITY_MIN, 1.0 - BP_SCORE_PROBABILITY_MIN)` expression with `probability`. | `bp_score_probability_clamps_saturated_llrs` |
 | `bp_runs_in_unpruned_fast_path` | In `FrontierDecoder::from_sparse_dem`, delete `&& !(config.k == usize::MAX && config.delta.is_infinite())` from the `bp_score` construction condition. | `bp_flag_is_bitwise_inert_on_the_unpruned_fast_path` (including its `bp_seconds == 0.0` assertion) |
-| `bptrellis_default_bp_off` | In `BpTrellisConfig::default`, replace `"bp_score_iterations: 5"` with `"bp_score_iterations: 0"`. | `bptrellis_defaults_enable_bp_merge_and_deadline_order` (including its `bp_seconds > 0.0` assertions) |
-| `bptrellis_default_merge_off` | In `BpTrellisConfig::default`, replace `"merge_indistinguishable: true"` with `"merge_indistinguishable: false"`. | `bptrellis_defaults_enable_bp_merge_and_deadline_order` (including its `processed_columns` assertion for the planted duplicate pair) |
-| `bptrellis_deadline_ignored` | In `BpTrellisDecoder::from_sparse_dem`, replace `"TrellisOrdering::Deadline => Some(deadline_column_order(dem)?)"` with `"TrellisOrdering::Deadline => None"`. | `bptrellis_defaults_enable_bp_merge_and_deadline_order`; `bptrellis_matches_hand_mapped_frontier_for_every_ordering` |
-| `escalate_on_any_failure` | In `BpTrellisDecoder::decode`, escalate a successful result whenever `result.status != FrontierStatus::Exact` instead of returning every `FrontierDecodeAttempt::Success` immediately. | `successful_base_decode_is_bit_identical_with_a_configured_ladder`; `wrong_prediction_does_not_escalate` |
-| `ladder_skips_accumulation` | Delete `"result.transitions += transitions;"` from the successful-rung branch in `BpTrellisDecoder::decode`. | `no_path_escalates_to_k16_and_accumulates_transitions` |
-| `ladder_reuses_base_k` | In `BpTrellisDecoder::from_sparse_dem`, replace `"k: rung_k"` with `"k"` when constructing each escalation rung. | `no_path_escalates_to_k16_and_accumulates_transitions` |
