+++ python/pecos-rslib-exp/Cargo.toml
@@ -25,6 +25,7 @@ extension-module = [
 ]
 
 [dependencies]
+pecos-bp-trellis.workspace = true
 pecos-core.workspace = true
 pecos-eeg.workspace = true
 pecos-frontier.workspace = true
